refactor: source code restructure
