+++ lib/auth/v2/auth.ts
