+++ lib/interfaces/http.ts
@@ -2,8 +2,8 @@ import express from 'express'
 import compression from 'compression'
 import cors from 'cors'
 
-import heyRouter from '../v1/hey/router'
-import userRouter from 'v1/users/router'
+import v1heyRouter from '../hey/v1/router'
+import v1userRouter from '../users/v1/router'
 
 const app = express()
 
@@ -13,7 +13,7 @@ app.use(compression())
 app.use(cors())
 app.disable('x-powered-by')
 
-app.use('/v1/hey', heyRouter)
-app.use('/v1/users', userRouter)
+app.use('/v1/hey', v1heyRouter)
+app.use('/v1/users', v1userRouter)
 
 export default app
