+++ package.json
@@ -53,9 +53,13 @@
 	},
 	"peerDependencies": {},
 	"dependencies": {
+		"aes-js": "^3.1.2",
+		"argon2": "^0.28.2",
+		"bcrypt": "^5.0.1",
 		"compression": "^1.7.4",
 		"consola": "^2.15.3",
 		"cors": "^2.8.5",
+		"crypto-js": "^4.1.1",
 		"csurf": "^1.11.0",
 		"dotenv": "^10.0.0",
 		"express": "^5.0.0-alpha.8",
